Dive summary: move implementation details away from header
Move all the details of the dive summary calculation into
the translation unit and only keep the global interface in
the header file.

Moreover, remove the diveSummary class, as it was only used
as a collection of functions anyway. We could think about
replacing it by a namespace, but a namespace for a single
function appears like overkill.

This commit is much smaller than it looks, because the
local functions were shuffled around to avoid having to
do declarations.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>