mobile-widgets: avoid spurious missing signal/slot handling
When connect() is in the constructor and the constructor constructs
a global object, there are no guarantee that the parallel object(s)
used in the connect() are constructed, which do not lead to an error
(something which Qt should catch, because connect() can check if
Q_OBJECT is initialized), but to a signal getting lost.

This is a somewhat random behaviour, which seems to occur more often
when debugging constructors of global objects. Recompiling often changes
the behaviour.

The effect of a missing signal is easily overlooked in the UI, because
it typically means a single graphical element is not updated, until
another event occurs.

Make setup() non-static to allow use of [this] as lambda.

Signed-off-by: jan Iversen <jan@casacondor.com>