documentation: remove reference to the building page on the Subsurface website
Removed the reference and link to the building page on the subsurface website
(https://subsurface-divelog.org/building) from CONTRIBUTIONG.md. This page no
longer exists and gives a 404 error. I have looked for suitable alternate
build/contributing instructions on the website and on github and can't find
anything so i think best to just remove the broken link for the time being. If
more comprehensive instructions exist in the future this will be simple to add
back in.

Signed-off-by: Jason Bramwell <jb2cool@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>