CylindersModel: test for CHANGED() outside of switch statement
A small code consolidation:

With one exception, all targets of the switch statement would
test for CHANGED(). Instead do the test once and exit early.
This changes the behavior of the function: if not changed, there
will be no more dataChanged-signal. However, this appears to be
the correct thing to do anyway. And it is easily changed if
it matters after all.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>