desktop: move undo-disabling from enterEditState to disableShortcuts
When entering the edit state, we don't want the user to be able
to undo/redo lest things become inconsistent. Since the only way
to enter edit state is to edit the profile, we can simply use
the disableShortcuts() function that is used by the profile
when it goes into edit state.

This has one desirable side-effect: Undo is now also disabled in
the planner. Undo during planning likewise can lead to inconsistent
state.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>