translations: update sources and translations
With the change in context for all the commands there was a lot thrash
in the translations. I tried to work around that as well as I could and
to keep the translations valid on Transifex, but a few appear to have
been lost in the shuffle.

But at least the command texts are now correctly translated.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>